First cut at getting things building with newer toolchains. - Default Mac SDK - Default compiler - 10.6 min for xibs and OS - Fix one compiler warning - Move to some newer apis to avoid dep. warnings. - Add a workspace into the project - Add some more svn:ignores into the tree.
git-svn-id: http://coverstory.googlecode.com/svn/trunk@166 c192c14b-6038-0410-9fcd-4bb9978d2467

+++ CoverStoryScroller.m
@@ -61,7 +61,7 @@
   // Draw our track
   OSStatus err = HIThemeDrawTrack(&drawInfo, NULL, context, 0);
   if (err) {
-    NSLog(@"Error drawing scroller: %d", err);
+    NSLog(@"Error drawing scroller: %d", (int)err);
   }
   
   // If we have coverage data, draw the lines to denote lines we didn't hit
